modules/cpuTemperature/requirements.txt
gpiozero==1.6.2
rpi.gpio
azure-iot-device~=2.7.0
asyncio